Big ol' update
Removed carousel and added hero section. Removed navbar on homepage since all the navigation to resources and concepts was redundant. Had to modify hash routing for it to accommodate the new grid, I think I fixed all the routing bugs but pay attention to the URLS just in case.

Placeholder image for the new support section on the homepage, deploy once that's updated.