CSS + ConceptSection further fixes
Modified display for ConceptSection items to be a flex display that stays stacked for sm and md, and displays side by side with figures on lg and above.

Fixed the issue with ConceptSection titles being overlapped by the navbar on md screens only. Pending article rewrites and some minor visual tune-ups post rewrite, this is more or less good to deploy after fully updating the character grid.